+++ src/pages/auth/google-callback/googleCallback.module.scss
@@ -1,0 +1,2 @@
+.GoogleCallback {
+}
